Altered file names
Files names changed for consistency
