+++ cleanUp.py
